get back the original tests
